Parquet: Refactor BaseParquetReaders
